Extract the delivery-control protocol into a leaf module; restore the size ratchet
Review round 1 (R1): loop.py crossed its shrink-only byte-debt entry
(319153 > 312765) and delegate_custody.py entered the giant band (1611 >
1600). The protocol-domain pieces — the DeliveryCandidate dataclass, the
hold-control vocabulary, the typed control prompt, the replace/keep
predicates, and the two pure parsers — move to the new leaf
ouroboros/delivery_protocol.py (never the ctx-touching resolvers); loop
re-imports them under the historical underscore names so callers, tests,
and monkeypatch targets keep one surface, and this sprint's own added
comments are tightened. delegate_custody's state= additions compress to
inline defaults (semantics and monkeypatch shapes unchanged), and the
reaper's audit import/call compress so reap_timed_out_task stays within
its function ceiling. Manifest regenerated: loop.py's BYTE_DEBT entry
shrinks to 312731 and the validator reports no issues.
