+++ ouroboros/size_ratchet_manifest.py
@@ -230,7 +230,7 @@ BYTE_BASELINE_DEBT = {
 }
 
 BYTE_DEBT = {
-    "ouroboros/loop.py": 312765,
+    "ouroboros/loop.py": 312731,
     "tests/test_delegated_subagent_transport.py": 320340,
     "tests/test_devtools_benchmarks.py": 328116,
     "web/modules/chat.js": 224244,
